.Se define BlockChainDataTypes.h para declarar los tipos de datos interfaz entre FileManager y Builder 	*Se modifica BlockChainBuilder, BlockChainFileManager y el 	makefile para incluir BlockChainDataTypes 	*Se corrije un error de escritura de state y status en FileManager

+++ Src/BlockChainBuilder.cpp
@@ -60,7 +60,7 @@ std::string BlockChainBuilder::Calculononce() {
 
 }
 
-bool BlockChainBuilder::CalculoBits( std::string hash, unsigned int bits ) {
+bool BlockChainBuilder::CalculoBits( std::string hash, size_t bits ) {
 	return true;
 }
